requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
phonenumbers==8.13.15
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.0
